start puma via bundler, fix typo
we install and manage puma via bundler (see our Gemfile). It is
possible, depending on the local bundler settings, that the puma binary
is only available for bundler and not via global $PATH. also the puma
bin provided by $PATH could be a wrong one. This commit ensures the
right puma version.

(also we fix a mini typo, a missing l)