implement rubocop as Rake task
to be future ready, we don't run rubocop by hand but through bundle ->
rake. This allows us to implement more tasks with different tests later
(unit/acceptance tests for example). I updated the docs and the
.travis.yml to refer to the rake task. Also fixed a wrong link in the
Makefile and a wrong listing in a readme that I found while testing this.