switch from inheritance to composition for node roles
as visible in our ERD, the three roles virt_node, ceph_osd_node and
ceph_mon_node are not hinherited from node, they compose the node. The
old implementation was based on inheritance, this commit fixes it.

* the roles inherit from ActiveRecord::Base as every model, not from node
* we defined the relation with belongs_to in the role models and their migrations
* updated the schema to the latest version
* fixed a typo in 006_initialize_virt.rb

updating existing schemas is a bad idea for stable software, however we
didn't hit 1.0.0 so far so this is okay.