introduce sinatra-contrib config handler
this commit provides a simple config handler which parses yaml files. We
also provide a basic yaml config with values that we later need for our
puppetdb/zabbix integration. gitignore gets extended to ignore the live
config because it can contain sensitive data (we also extend the doc in
virtapi_app.rb)